Move event fabric classes into the existing internal/api/event package
